requirements.txt
argparse==1.1
numpy==1.24.2
biopython==1.81
matplotlib==3.5.1
